Move CredentialConfigMapper to util package
